batch/requirements.txt
tqdm
langchain
pysolr
kazoo
sentence-transformers
fugashi
ipadic
mysql-connector-python
pymongo
eurelis-langchain-solr-vectorstore
pandas
